requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.36
langchain-groq==0.1.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
PyMuPDF==1.24.0
Pillow==10.3.0
cryptography==42.0.5
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4
python-dotenv==1.0.1
pandas==2.2.1
plotly==5.20.0